+++ app_note.kt
@@ -1,4 +1,4 @@
-// https://youtu.be/ERBEWmfz6h0?si=0TexLj0_8gM3Xghg
+// https://www.youtube.com/watch?v=ERBEWmfz6h0&list=PLSrm9z4zp4mEWwyiuYgVMWcDFdsebhM-r
 
 /*
 Mian link : https://www.youtube.com/watch?v=ERBEWmfz6h0&list=PLSrm9z4zp4mEWwyiuYgVMWcDFdsebhM-r
@@ -17,6 +17,7 @@ Next  video Box .
 
 https://youtu.be/iIm-7fmjyaA?si=KIJ1jo6odnrFssq_
 
+Next : Text Selection - Jetpack Compose
 
 
 *
